增加recipe下载接口和recipe下载日志生成方法

+++ lmrj-modules/lmrj-module-rms/src/main/java/com/lmrj/rms/log/entity/RmsRecipeLog.java
@@ -45,12 +45,12 @@ public class RmsRecipeLog extends BaseDataEntity {
     private String versionType;
     /**配方版本号*/
     @TableField(value = "version_no")
-    private Integer versionNo;
+    private Double versionNo;
     /**批次id*/
     @TableField(value = "lot_id")
     private String lotId;
     /**参数信息*/
     @TableField(value = "param")
     private String param;
-	
+
 }
